unwrap: Change 3D wrapper code from float to double.

+++ skimage/exposure/_unwrap_3d.pyx
@@ -1,12 +1,12 @@
-cdef extern void unwrap3D(float* wrapped_volume,
-                     float* unwrapped_volume,
+cdef extern void unwrap3D(double* wrapped_volume,
+                     double* unwrapped_volume,
                      unsigned char* input_mask,
                      int image_width, int image_height, int volume_depth,
                      int wrap_around_x, int wrap_around_y, int wrap_around_z)
 
-def unwrap_3d(float[:, :, ::1] image,
+def unwrap_3d(double[:, :, ::1] image,
               unsigned char[:, :, ::1] mask,
-              float[:, :, ::1] unwrapped_image,
+              double[:, :, ::1] unwrapped_image,
               wrap_around):
     unwrap3D(&image[0, 0, 0],
              &unwrapped_image[0, 0, 0],
